1_Download modified, Added dependencies,RSetup, Tests
I modified the first Script: 1_Downdload_seqs.R so it contains only the function that can be sourced in any other script. I also added the Dependencies.R script just to follow what are the dependencies of the whole workflow. This could be a simple text file... RSetup is a function to install, load libraries and set working dir ideally the directory where the data and other files of a particular analyses should be placed. And finally the Tests script can be used to test each of the functions.